requirements.txt
requests
pypdf
langchain
python-decouple
python-dotenv
playwright
streamlit
langchain-community
langchain-openai
langchain-core
langchain-pinecone
nest-asyncio
lxml 
better_profanity
langgraph
ipython
ipykernel
streamlit-feedback
redis
langchain-redis 
cassio
singlestoredb
pymysql